+++ ModernGL/renderbuffers.py
@@ -38,6 +38,12 @@ class Renderbuffer:
     def __repr__(self):
         return '<Renderbuffer: %d>' % self.glo
 
+    def __eq__(self, other):
+        return self.mglo is other.mglo
+
+    def __ne__(self, other):
+        return self.mglo is not other.mglo
+
     @property
     def width(self) -> int:
         '''