+++ src/BufferAccess.cpp
@@ -55,12 +55,12 @@ PyObject * MGLBufferAccess_close(MGLBufferAccess * self) {
 }
 
 PyObject * MGLBufferAccess_read(MGLBufferAccess * self, PyObject * args) {
-	int size;
-	int offset;
+	Py_ssize_t size;
+	Py_ssize_t offset;
 
 	int args_ok = PyArg_ParseTuple(
 		args,
-		"II",
+		"nn",
 		&size,
 		&offset
 	);
@@ -69,11 +69,11 @@ PyObject * MGLBufferAccess_read(MGLBufferAccess * self, PyObject * args) {
 		return 0;
 	}
 
-	if (size == -1) {
+	if (size < 0) {
 		size = self->size - offset;
 	}
 
-	if (offset < 0 || size < 0 || size + offset > self->size) {
+	if (offset < 0 || offset + size > self->size) {
 		MGLError_Set("out of range offset = %d or size = %d", offset, size);
 		return 0;
 	}
@@ -88,13 +88,13 @@ PyObject * MGLBufferAccess_read(MGLBufferAccess * self, PyObject * args) {
 
 PyObject * MGLBufferAccess_read_into(MGLBufferAccess * self, PyObject * args) {
 	PyObject * data;
-	int size;
-	int offset;
-	int write_offset;
+	Py_ssize_t size;
+	Py_ssize_t offset;
+	Py_ssize_t write_offset;
 
 	int args_ok = PyArg_ParseTuple(
 		args,
-		"OIII",
+		"Onnn",
 		&data,
 		&size,
 		&offset,
@@ -105,11 +105,11 @@ PyObject * MGLBufferAccess_read_into(MGLBufferAccess * self, PyObject * args) {
 		return 0;
 	}
 
-	if (size == -1) {
+	if (size < 0) {
 		size = self->size - offset;
 	}
 
-	if (offset < 0 || size < 0 || size + offset > self->size) {
+	if (offset < 0 || offset + size > self->size) {
 		MGLError_Set("out of range offset = %d or size = %d", offset, size);
 		return 0;
 	}
@@ -142,12 +142,12 @@ PyObject * MGLBufferAccess_read_into(MGLBufferAccess * self, PyObject * args) {
 
 PyObject * MGLBufferAccess_write(MGLBufferAccess * self, PyObject * args) {
 	const char * data;
-	int size;
-	int offset;
+	Py_ssize_t size;
+	Py_ssize_t offset;
 
 	int args_ok = PyArg_ParseTuple(
 		args,
-		"y#I",
+		"y#n",   // TODO: replace y# with O if possible
 		&data,
 		&size,
 		&offset
@@ -182,11 +182,11 @@ PyMethodDef MGLBufferAccess_tp_methods[] = {
 };
 
 PyObject * MGLBufferAccess_get_offset(MGLBufferAccess * self) {
-	return PyLong_FromLong(self->offset);
+	return PyLong_FromSsize_t(self->offset);
 }
 
 PyObject * MGLBufferAccess_get_size(MGLBufferAccess * self) {
-	return PyLong_FromLong(self->size);
+	return PyLong_FromSsize_t(self->size);
 }
 
 PyObject * MGLBufferAccess_get_readonly(MGLBufferAccess * self) {